fix: address review comments
- Remove unpublish from NOT_IMPLEMENTED list so the real command wins
- Extract shared parsePackageSpec and rcOptionsTypes into common.ts
  to deduplicate code between deprecate and unpublish
