+++ registry-access/commands/src/deprecation/common.ts
@@ -1,19 +1,14 @@
 import { pickRegistryForPackage } from '@pnpm/config.pick-registry-for-package'
-import { types as allTypes } from '@pnpm/config.reader'
 import { PnpmError } from '@pnpm/error'
 import { createGetAuthHeaderByURI } from '@pnpm/network.auth-header'
 import { createFetchFromRegistry, type CreateFetchFromRegistryOptions } from '@pnpm/network.fetch'
-import npa from '@pnpm/npm-package-arg'
 import type { PackageInRegistry, PackageMeta } from '@pnpm/resolving.registry.types'
 import type { Registries, RegistryConfig } from '@pnpm/types'
-import { pick } from 'ramda'
 import semver from 'semver'
 
-export function rcOptionsTypes (): Record<string, unknown> {
-  return pick([
-    'registry',
-  ], allTypes)
-}
+import { parsePackageSpec, rcOptionsTypes } from '../common.js'
+
+export { parsePackageSpec, rcOptionsTypes }
 
 export function cliOptionsTypes (): Record<string, unknown> {
   return {
@@ -28,20 +23,6 @@ export interface DeprecateOptions extends CreateFetchFromRegistryOptions {
   }
   configByUri?: Record<string, RegistryConfig>
   registries?: Registries
-}
-
-export function parsePackageSpec (spec: string): { name: string, versionRange: string | undefined } {
-  let parsed: ReturnType<typeof npa>
-  try {
-    parsed = npa(spec)
-  } catch {
-    throw new PnpmError('INVALID_PACKAGE_SPEC', `Invalid package spec: ${spec}`)
-  }
-  if (!parsed.name) {
-    throw new PnpmError('INVALID_PACKAGE_SPEC', `Invalid package spec: ${spec}`)
-  }
-  const versionRange = parsed.rawSpec || undefined
-  return { name: parsed.name, versionRange }
 }
 
 interface UpdateDeprecationOptions {
